angular refactoring moved all provider components to provider folder
