+++ frontend-distribution-platform/src/app/app.module.ts
@@ -39,7 +39,7 @@ import {
 } from '@nebular/theme';
 import { RegisterProviderComponent } from './register-provider/register-provider.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { ProviderServiceListComponent } from './provider-service-list/provider-service-list.component';
+import { ProviderServiceListComponent } from './provider/provider-service-list/provider-service-list.component';
 import { BusinessStatsComponent } from './business/business-stats/business-stats.component';
 import { BusinessStatsProgressSection } from './business/business-stats/progress-section/progress-section.component';
 import { BusinessStatsCharts } from './business/business-stats/charts-panel/charts-panel.component';
@@ -53,17 +53,17 @@ import { NgxEchartsModule } from 'ngx-echarts';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
 import { RegisterBusinessComponent } from './register-business/register-business.component';
-import { ProviderServiceFormComponent } from './provider-service-form/provider-service-form.component';
-import { ProviderProfileComponent } from './provider-profile/provider-profile.component';
-import { ProviderRequestsListComponent } from './provider-requests-list /provider-requests-list.component';
-import { ProviderStatsComponent } from './provider-stats/provider-stats.component';
-import { ProviderStatsCharts} from './provider-stats/charts-panel/charts-panel.component';
-import { ProviderStatsProgressSection } from './provider-stats/progress-section/progress-section.component';
-import { ProviderStatsChartsProfitChartComponent } from './provider-stats/charts-panel/charts/profit-chart.component';
-import { ProviderStatsChartsOrdersChartComponent } from './provider-stats/charts-panel/charts/orders-chart.component';
-import { ProviderStatsChartsChartPanelSummaryComponent } from './provider-stats/charts-panel/chart-panel-summary/chart-panel-summary.component';
-import { ProviderStatsChartsChartPanelHeaderComponent } from './provider-stats/charts-panel/chart-panel-header/chart-panel-header.component';
-import { ProviderStatsChartsLegendChartComponent } from './provider-stats/charts-panel/legend-chart/legend-chart.component';
+import { ProviderServiceFormComponent } from './provider/provider-service-form/provider-service-form.component';
+import { ProviderProfileComponent } from './provider/provider-profile/provider-profile.component';
+import { ProviderRequestsListComponent } from './provider/provider-requests-list /provider-requests-list.component';
+import { ProviderStatsComponent } from './provider/provider-stats/provider-stats.component';
+import { ProviderStatsCharts} from './provider/provider-stats/charts-panel/charts-panel.component';
+import { ProviderStatsProgressSection } from './provider/provider-stats/progress-section/progress-section.component';
+import { ProviderStatsChartsProfitChartComponent } from './provider/provider-stats/charts-panel/charts/profit-chart.component';
+import { ProviderStatsChartsOrdersChartComponent } from './provider/provider-stats/charts-panel/charts/orders-chart.component';
+import { ProviderStatsChartsChartPanelSummaryComponent } from './provider/provider-stats/charts-panel/chart-panel-summary/chart-panel-summary.component';
+import { ProviderStatsChartsChartPanelHeaderComponent } from './provider/provider-stats/charts-panel/chart-panel-header/chart-panel-header.component';
+import { ProviderStatsChartsLegendChartComponent } from './provider/provider-stats/charts-panel/legend-chart/legend-chart.component';
 import { BusinessProfileComponent } from './business/business-profile/business-profile.component';
 import { BusinessAPIComponent } from './business/business-api/business-api.component';
 import { SharedModule } from './shared/shared.module';
