+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/chart-panel-header/chart-panel-header.component.html
