+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/chart-panel-header/chart-panel-header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../@theme/styles/themes';
+@import '../../../../@theme/styles/themes';
 @import '~bootstrap/scss/mixins/breakpoints';
 @import '~@nebular/theme/styles/global/breakpoints';
