+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/chart-panel-summary/chart-panel-summary.component.scss
@@ -1,4 +1,4 @@
-@import '../../../@theme/styles/themes';
+@import '../../../../@theme/styles/themes';
 
 @include nb-install-component() {
   .summary-container {
