+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/chart-panel-summary/chart-panel-summary.component.ts
