+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/charts-panel.component.html
