+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/charts-panel.component.scss
@@ -1,4 +1,4 @@
-@import '../../@theme/styles/themes';
+@import '../../../@theme/styles/themes';
 
 $legend-all-orders-color: #00977e;
 $legend-payment-color: #6935ca;
