+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/charts/charts-common.component.scss
@@ -1,4 +1,4 @@
-@import '../../../@theme/styles/themes';
+@import '../../../../@theme/styles/themes';
 
 @include nb-install-component() {
   flex: 1;
