+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/charts/orders-chart.component.ts
@@ -2,8 +2,8 @@ import { AfterViewInit, Component, Input, OnChanges, OnDestroy } from '@angular/
 import { NbThemeService } from '@nebular/theme';
 import { delay, takeWhile } from 'rxjs/operators';
 
-import { OrdersChart } from '../../../@core/data/orders-chart';
-import { LayoutService } from '../../../@core/utils/layout.service';
+import { OrdersChart } from '../../../../@core/data/orders-chart';
+import { LayoutService } from '../../../../@core/utils/layout.service';
 
 @Component({
   selector: 'ngx-provider-stats-charts-orders-chart',
