+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/charts/profit-chart.component.ts
@@ -2,8 +2,8 @@ import { AfterViewInit, Component, Input, OnChanges, OnDestroy } from '@angular/
 import { NbThemeService } from '@nebular/theme';
 import { takeWhile } from 'rxjs/operators';
 
-import { ProfitChart } from '../../../@core/data/profit-chart';
-import { LayoutService } from '../../../@core/utils/layout.service';
+import { ProfitChart } from '../../../../@core/data/profit-chart';
+import { LayoutService } from '../../../../@core/utils/layout.service';
 
 @Component({
   selector: 'ngx-provider-stats-charts-profit-chart',
