+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/legend-chart/enum.legend-item-color.ts
