+++ frontend-distribution-platform/src/app/provider/provider-stats/charts-panel/legend-chart/legend-chart.component.ts
