+++ frontend-distribution-platform/src/app/provider/provider-stats/progress-section/progress-section.component.html
