+++ frontend-distribution-platform/src/app/provider/provider-stats/progress-section/progress-section.component.scss
@@ -1,4 +1,4 @@
-@import '../../@theme/styles/themes';
+@import '../../../@theme/styles/themes';
 
 $shadow-green: #00977e;
